// swe642-HW4/src/app/models/StudentSurvey.model.ts
//Model to represent a student survey
export class StudentSurvey {
  name: string;
  studentId: string;
  address: string;
  city: string;
  state: string;
  zip: number;
  telephone: string;
  email: string;
  date: string;
  url: string;
  data: string;
  gradMonth: string;
  gradYear: number;
  preferences: string;
  interested: string;
  recommend: string;
  comments: string
}